Renamed the original UnitTestSiloHost base class, moved members to utility classes and moved the HostedCluster classes to a separate code file
