+++ src/TesterInternal/StorageTests/AzureQueueDataManagerTests.cs
@@ -8,6 +8,7 @@ using Orleans.Runtime;
 using Orleans.Runtime.Configuration;
 using Orleans.AzureUtils;
 using Orleans.TestingHost;
+using Tester;
 using UnitTests.Tester;
 
 namespace UnitTests.StorageTests
@@ -30,7 +31,7 @@ namespace UnitTests.StorageTests
         [ClassInitialize]
         public static void ClassInitialize(TestContext testContext)
         {
-            UnitTestSiloHost.CheckForAzureStorage();
+            TestUtils.CheckForAzureStorage();
         }
 
         [TestCleanup]
